+++ Remnants/Behaviours/BodyMovementBehaviour.cs
@@ -1,11 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
+﻿using System.Linq;
 using UnityEngine;
-using UnityEngine.SceneManagement;
 
 namespace Remnants.Behaviours
 {