+++ Remnants/Behaviours/RegisterBodiesSpawnBehaviour.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using UnityEngine.SceneManagement;
 using UnityEngine;
 using LethalLib.Modules;
-using Unity.Netcode;
 using BepInEx;
 
 namespace Remnants.Behaviours
@@ -47,17 +46,7 @@ namespace Remnants.Behaviours
 
             mls.LogInfo("Registering bodies to moons");
             _isRegistering = true;
-            List<string> planetNames = new List<string>();
-            foreach (var level in startOfRound.levels)
-            {
-                if (HasIllegalCharacters(level.PlanetName) || PlanetsBodiesRarities.ContainsKey(level.PlanetName))
-                    continue;
-
-                RegisterBodiesToNewMoon(level);
-                if(!Enum.TryParse(level.name, out Levels.LevelTypes a))
-                    planetNames.Add(level.PlanetName);
-            }
-            Data.Config.SetCustomLevelsRarities(planetNames);
+            RegisterData(startOfRound.levels);
             SceneManager.sceneLoaded -= RegisterBodiesToMoons;
             _isRegistering = false;
         }
@@ -96,6 +85,20 @@ namespace Remnants.Behaviours
             return name.IndexOfAny(chars) != -1;
         }
 
+        private void RegisterData(SelectableLevel[] levels)
+        {
+            List<string> planetNames = new List<string>();
+            foreach (var level in levels)
+            {
+                if (HasIllegalCharacters(level.PlanetName) || PlanetsBodiesRarities.ContainsKey(level.PlanetName))
+                    continue;
+
+                RegisterBodiesToNewMoon(level);
+                if (!Enum.TryParse(level.name, out Levels.LevelTypes a))
+                    planetNames.Add(level.PlanetName);
+            }
+            Data.Config.SetCustomLevelsRarities(planetNames);
+        }
         #endregion
     }
 }