+++ Remnants/Data/LoadAssetsBodies.cs
@@ -4,7 +4,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
-using Unity.Netcode;
 using UnityEngine;
 using UnityEngine.SceneManagement;