+++ Remnants/Main.cs
@@ -18,7 +18,7 @@ namespace Remnants
         #region Variables
         private const string modGUID = "KawaiiBone.Remnants";
         private const string modName = "Remnants";
-        private const string modVersion = "1.2.7";
+        private const string modVersion = "1.2.8";
 
         public static Remnants Instance;
         private readonly Harmony _harmony = new Harmony(modGUID);